+++ tests/PSRule.Tests/PSRule.Variables.Tests.ps1
@@ -24,6 +24,8 @@ $here = (Resolve-Path $PSScriptRoot).Path;
 #region PSRule variables
 
 Describe 'PSRule variables' -Tag 'Variables' {
+    $ruleFilePath = (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1');
+
     Context 'PowerShell automatic variables' {
         $testObject = [PSCustomObject]@{
             Name = 'VariableTest'
@@ -32,10 +34,25 @@ Describe 'PSRule variables' -Tag 'Variables' {
         $testObject.PSObject.TypeNames.Insert(0, $testObject.Type);
 
         It '$Rule' {
-            $result = $testObject | Invoke-PSRule -Path (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1') -Name 'VariableTest';
+            $result = $testObject | Invoke-PSRule -Path $ruleFilePath -Name 'VariableTest';
             $result | Should -Not -BeNullOrEmpty;
             $result.IsSuccess() | Should -Be $True;
             $result.TargetName | Should -Be 'VariableTest';
+        }
+
+        It '$LocalizedData' {
+            $invokeParams = @{
+                Path = $ruleFilePath
+                Name = 'WithLocalizedData'
+                Culture = 'en-ZZ'
+                WarningAction = 'SilentlyContinue'
+            }
+            $result = $testObject | Invoke-PSRule @invokeParams -WarningVariable outWarning;
+            $messages = @($outwarning);
+            $result | Should -Not -BeNullOrEmpty;
+            $result.IsSuccess() | Should -Be $True;
+            $result.TargetName | Should -Be 'VariableTest';
+            $messages[0] | Should -Be 'LocalizedMessage for en-ZZ. Format=TestType.';
         }
     }
 }